+++ src/cherry_etl/__init__.py
@@ -1,4 +1,5 @@
 from . import config
 from .pipeline import run_pipeline
+from . import datasets
 
-__all__ = ["config", "run_pipeline"]
+__all__ = ["config", "run_pipeline", "datasets"]
