+++ src/cherry_etl/datasets/__init__.py
@@ -1,0 +1,3 @@
+from . import evm, svm
+
+__all__ = ["evm", "svm"]
