+++ src/cherry_etl/datasets/evm/__init__.py
@@ -1,0 +1,3 @@
+from .address_appearances import make_pipeline as address_appearances
+
+__all__ = ["address_appearances"]
